+++ prr-core/src/prr/plans/NormalPlan.java
@@ -8,27 +8,27 @@ public class NormalPlan implements Plan {
     private static final long serialVersionUID = 202208091754L;
 
     @Override
-    public long calculateMessageCost(int units){
+    public double calculateMessageCost(double units){
         if (units < 50)
             return 10;
         if (units < 100)
             return 16;
-        return (long) 2*units;
+        return 2*units;
     }
 
     @Override
-    public long calculateVoiceCommunicationCost(int units, boolean isFriend){
-        long cost = (long) 20*units;
+    public double calculateVoiceCommunicationCost(double units, boolean isFriend){
+        double cost = 20*units;
         if (isFriend)
-            return (long) (cost*0.5);
+            return cost*0.5;
         return cost;
     }
 
     @Override
-    public long calculateVideoCommunicationCost(int units, boolean isFriend){
-        long cost = (long) 30*units;
+    public double calculateVideoCommunicationCost(double units, boolean isFriend){
+        double cost = 30*units;
         if (isFriend)
-            return (long) (cost*0.5);
+            return cost*0.5;
         return cost;
     }
