+++ prr-core/src/prr/terminals/communication/Video.java
@@ -15,7 +15,7 @@ public class Video extends InteractiveCommunication implements Serializable{
         super(source, destination, communicationId);
     }
 
-    protected long calculateCost(int units){
+    protected double calculateCost(double units){
         Terminal source = getSource();
         boolean isFriend = source.isFriend(getDestination().getKey());
         return source.getOwner().getPlan().calculateVideoCommunicationCost(units, isFriend);
